Show procced button when video has no source
This fixes to problem with the check for valid video.
The first one is that the old check was for invalid video and would return that the video wasn't invalid even if it
wasn't a video. The new check make sure it is a video and has valid sources.
The other problem was that it old library was sent in to the check. Now it is the one that will shown that is checked for a valid video.